Finished moving timeslot methods to timeslot controller
